move wizard for saving to wizard package and add javadoc
