+++ src/main/resources/wizard_save_file.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.*?>
 
 <VBox fx:id="root" styleClass="main-box-save" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-	  fx:controller="controllers.SaveWizardController">
+	  fx:controller="controllers.wizard.SaveWizardController">
 	<children>
 		<HBox styleClass="save-box">
 			<!--main top box-->
